+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+### nova-dashboard, 0.23.0
+
+* Display instrument mount alerts specific to each available compute node (thanks to John Duggan).
+* Warn users when running jobs while compute nodes or instrument mounts are experiencing outages (thanks to John Duggan).
+
 ### nova-dashboard, 0.22.0
 
 * Add tool documentation links (thanks to John Duggan).